requirements.txt
fastapi
uvicorn[standard]
pydantic
asyncpg
databases[asyncpg]
boto3
python-dotenv
sqlalchemy
psycopg2-binary